상세페이지 스타일 수정
